+++ src/Components/Comment/index.js
@@ -1,0 +1,1 @@
+export { default as CommentInput } from "./CommentInput";
