+++ src/Components/NavItem/index.js
@@ -1,0 +1,1 @@
+export { default as NavItem } from "./NavItem";
