+++ src/Components/TocItem/index.js
@@ -1,0 +1,1 @@
+export { default as TocItem } from "./TocItem";
