+++ src/Components/Toggle/index.jsx
@@ -1,0 +1,1 @@
+export { default as Toggle } from "./Toggle";
