+++ src/Foundations/Avatar/index.js
@@ -1,0 +1,2 @@
+export { default as Avatar } from "./Avatar";
+export { default as UserAlt } from "./UserAlt";
