+++ src/Foundations/Icon/index.js
@@ -1,0 +1,3 @@
+export { default as Icon } from "./Icon";
+export { default as IconBookmark } from "./Icon";
+export { default as IconSocial } from "./Icon";
