+++ src/Foundations/Logo/index.js
@@ -1,0 +1,1 @@
+export { default as Logo } from "./Logo";
