--- src/Logo/Logo.js
@@ -1,90 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-import styled from "styled-components";
-import { color } from "../shared/styles";
-
-const THEME = {
-  DARK: "dark",
-  BLUE: "blue",
-  LIGHT: "light",
-  TRANSPARENT: "transparent",
-};
-
-export const Logo = ({ width, height, theme }) => {
-  return (
-    <Svg
-      width={width}
-      height={height}
-      viewBox="0 0 1000 260"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g>
-        <Symbol
-          d="M27.6355 243.543C-7.37624 212.545 -9.7667 144.382 24.1665 83.1562C57.3834 23.2187 114.155 -9.749 157.915 2.5626C47.0851 44.0717 139.436 202.596 27.6355 243.543Z"
-          theme={theme}
-        />
-        <Symbol
-          d="M186.833 176.846C151.338 240.886 88.9612 274.137 44.291 254.269C158.671 213.777 63.3986 52.5323 177.875 12.0898C218.042 40.6145 222.461 112.556 186.833 176.846Z"
-          theme={theme}
-        />
-      </g>
-      <Wordmark
-        d="M393.64 213.98C381.869 213.98 371.31 211.9 361.963 207.74C352.615 203.407 345.172 197.6 339.633 190.32C334.267 183.04 331.324 174.98 330.805 166.14H367.415C368.108 171.687 370.791 176.28 375.464 179.92C380.311 183.56 386.283 185.38 393.38 185.38C400.304 185.38 405.67 183.993 409.479 181.22C413.46 178.447 415.45 174.893 415.45 170.56C415.45 165.88 413.027 162.413 408.18 160.16C403.507 157.733 395.977 155.133 385.591 152.36C374.859 149.76 366.03 147.073 359.106 144.3C352.356 141.527 346.47 137.28 341.45 131.56C336.604 125.84 334.18 118.127 334.18 108.42C334.18 100.447 336.43 93.1667 340.931 86.58C345.605 79.9934 352.183 74.7934 360.664 70.98C369.319 67.1667 379.446 65.26 391.043 65.26C408.18 65.26 421.855 69.5934 432.068 78.26C442.281 86.7534 447.907 98.28 448.945 112.84H414.152C413.633 107.12 411.21 102.613 406.882 99.32C402.728 95.8534 397.102 94.12 390.005 94.12C383.427 94.12 378.321 95.3334 374.685 97.76C371.223 100.187 369.492 103.567 369.492 107.9C369.492 112.753 371.916 116.48 376.763 119.08C381.609 121.507 389.139 124.02 399.352 126.62C409.738 129.22 418.307 131.907 425.057 134.68C431.808 137.453 437.607 141.787 442.454 147.68C447.474 153.4 450.07 161.027 450.243 170.56C450.243 178.88 447.907 186.333 443.233 192.92C438.732 199.507 432.155 204.707 423.5 208.52C415.018 212.16 405.064 213.98 393.64 213.98Z"
-        theme={theme}
-      />
-      <Wordmark
-        d="M529.936 213.98C518.165 213.98 507.606 211.9 498.258 207.74C488.911 203.407 481.468 197.6 475.928 190.32C470.562 183.04 467.62 174.98 467.1 166.14H503.711C504.403 171.687 507.086 176.28 511.76 179.92C516.607 183.56 522.579 185.38 529.676 185.38C536.6 185.38 541.966 183.993 545.774 181.22C549.756 178.447 551.746 174.893 551.746 170.56C551.746 165.88 549.323 162.413 544.476 160.16C539.802 157.733 532.272 155.133 521.886 152.36C511.154 149.76 502.326 147.073 495.402 144.3C488.651 141.527 482.766 137.28 477.746 131.56C472.899 125.84 470.476 118.127 470.476 108.42C470.476 100.447 472.726 93.1667 477.227 86.58C481.9 79.9934 488.478 74.7934 496.96 70.98C505.615 67.1667 515.741 65.26 527.339 65.26C544.476 65.26 558.151 69.5934 568.364 78.26C578.577 86.7534 584.202 98.28 585.241 112.84H550.448C549.929 107.12 547.505 102.613 543.178 99.32C539.023 95.8534 533.398 94.12 526.301 94.12C519.723 94.12 514.616 95.3334 510.981 97.76C507.519 100.187 505.788 103.567 505.788 107.9C505.788 112.753 508.212 116.48 513.058 119.08C517.905 121.507 525.435 124.02 535.648 126.62C546.034 129.22 554.602 131.907 561.353 134.68C568.104 137.453 573.903 141.787 578.75 147.68C583.77 153.4 586.366 161.027 586.539 170.56C586.539 178.88 584.202 186.333 579.529 192.92C575.028 199.507 568.45 204.707 559.795 208.52C551.313 212.16 541.36 213.98 529.936 213.98Z"
-        theme={theme}
-      />
-      <Wordmark
-        d="M601.838 139.1C601.838 124.54 604.694 111.627 610.407 100.36C616.292 89.0934 624.168 80.4267 634.035 74.36C644.075 68.2934 655.24 65.26 667.53 65.26C678.262 65.26 687.609 67.4267 695.572 71.76C703.707 76.0934 710.199 81.5534 715.045 88.14V67.6H751.656V211.64H715.045V190.58C710.372 197.34 703.881 202.973 695.572 207.48C687.436 211.813 678.002 213.98 667.27 213.98C655.153 213.98 644.075 210.86 634.035 204.62C624.168 198.38 616.292 189.627 610.407 178.36C604.694 166.92 601.838 153.833 601.838 139.1ZM715.045 139.62C715.045 130.78 713.314 123.24 709.852 117C706.39 110.587 701.717 105.733 695.831 102.44C689.946 98.9734 683.628 97.24 676.877 97.24C670.126 97.24 663.895 98.8867 658.182 102.18C652.47 105.473 647.796 110.327 644.161 116.74C640.699 122.98 638.968 130.433 638.968 139.1C638.968 147.767 640.699 155.393 644.161 161.98C647.796 168.393 652.47 173.333 658.182 176.8C664.068 180.267 670.299 182 676.877 182C683.628 182 689.946 180.353 695.831 177.06C701.717 173.593 706.39 168.74 709.852 162.5C713.314 156.087 715.045 148.46 715.045 139.62Z"
-        theme={theme}
-      />
-      <Wordmark
-        d="M847.917 97.5H822.731V211.64H785.861V97.5H769.503V67.6H785.861V60.32C785.861 42.64 790.881 29.64 800.921 21.32C810.96 13 826.107 9.1 846.359 9.62V40.3C837.531 40.1267 831.386 41.6 827.924 44.72C824.462 47.84 822.731 53.4734 822.731 61.62V67.6H847.917V97.5Z"
-        theme={theme}
-      />
-      <Wordmark
-        d="M1000 136.5C1000 141.7 999.655 146.38 998.962 150.54H893.804C894.67 160.94 898.305 169.087 904.709 174.98C911.114 180.873 918.99 183.82 928.338 183.82C941.839 183.82 951.446 178.013 957.159 166.4H996.366C992.211 180.267 984.249 191.707 972.478 200.72C960.707 209.56 946.253 213.98 929.117 213.98C915.269 213.98 902.805 210.947 891.727 204.88C880.822 198.64 872.253 189.887 866.022 178.62C859.963 167.353 856.934 154.353 856.934 139.62C856.934 124.713 859.963 111.627 866.022 100.36C872.08 89.0934 880.562 80.4267 891.467 74.36C902.373 68.2934 914.922 65.26 929.117 65.26C942.791 65.26 954.995 68.2067 965.727 74.1C976.632 79.9934 985.028 88.4 990.913 99.32C996.972 110.067 1000 122.46 1000 136.5ZM962.352 126.1C962.179 116.74 958.803 109.287 952.225 103.74C945.648 98.02 937.598 95.16 928.078 95.16C919.077 95.16 911.46 97.9334 905.229 103.48C899.17 108.853 895.449 116.393 894.064 126.1H962.352ZM956.899 29.64L897.18 56.94V30.94L956.899 0V29.64Z"
-        theme={theme}
-      />
-    </Svg>
-  );
-};
-
-Logo.propTypes = {
-  theme: PropTypes.oneOf(Object.values(THEME)),
-};
-
-Logo.defaultProps = {
-  theme: THEME.LIGHT,
-  width: 1000,
-  height: 260,
-};
-
-const symbolColors = {
-  dark: color.blue100,
-  blue: color.white,
-  light: color.blue100,
-  transparent: color.blue100,
-};
-
-const wordColors = {
-  dark: color.gray25,
-  blue: color.white,
-  light: color.gray900,
-  transparent: color.gray25,
-};
-
-const Svg = styled.svg`
-  display: block;
-`;
-
-const Symbol = styled.path`
-  fill: ${(props) => symbolColors[props.theme]};
-`;
-
-const Wordmark = styled.path`
-  fill: ${(props) => wordColors[props.theme]};
-`;
